Run `make` during `prepare` phase
Currently is not possible to install package via npm install
jcoglan/canopy because lib/ folders is not generated by npm. It makes
the local installation is not trivial process especially when version
deployed to npm is broken due to StringIndexOutOfBoundsException and
only master version is working fine.

This PR adds `make` to `prepare` phase.